pyproject.toml
[project]
name = "play-motion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mediapipe>=0.10.33",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "pyaudio>=0.2.14",
    "pygame>=2.6.1",
    "vosk>=0.3.45",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.9",
]
